refactor(core): split database/core.rs into health.rs, tests.rs modules (Phase 3, closes #189)
- Extract health/metrics impl block (is_connected, get_pool_health,
  get_pool_metrics, comprehensive_health_check, get_stats) into
  database/health.rs (~144 LOC)
- Move apply_sqlite_optimizations to pool.rs as pub(crate) free fn;
  make config field pub(crate) to allow sibling access
- Relocate #[cfg(test)] mod tests to database/tests.rs (~1028 LOC);
  declared from database/mod.rs so super::* resolves correctly
- Remove empty TAG/Bulk placeholder comment blocks from core.rs
- core.rs reduced from 1,425 → 163 lines (struct + constructors only)
