+++ libs/things3-core/src/database/mod.rs
@@ -2,6 +2,7 @@
 
 pub(crate) mod conversions;
 mod core;
+mod health;
 pub mod date_utils;
 pub mod mappers;
 mod mutations;
@@ -12,6 +13,8 @@ pub mod query_builders;
 pub(crate) mod stats;
 pub mod tag_utils;
 pub mod validators;
+#[cfg(test)]
+mod tests;
 
 // Re-export everything from core for backward compatibility
 pub use core::*;
